requirements.txt
# AI-Based Acne Detection & Skincare Recommendation - Dependencies
# Python 3.8+ recommended

tensorflow>=2.10,<2.16
Pillow>=9.0
numpy>=1.21
pandas>=1.3
matplotlib>=3.5
scikit-learn>=1.0
seaborn>=0.12
streamlit>=1.20
openai>=1.0.0
python-dotenv>=1.0.0
